Time: 10 ms (66.82%), Space: 15.9 MB (33.07%) - LeetHub

+++ 0229-majority-element-ii/0229-majority-element-ii.cpp
@@ -8,7 +8,7 @@ public:
     map<int, int> mpp;
 
     // least occurrence of the majority element:
-    int mini = int(n / 3) + 1;
+    int mini = (n / 3) + 1;
 
     //storing the elements with its occurnce:
     for (int i = 0; i < n; i++) {
